Add "fix" for older servers using API 4.1-BETA
