+++ src/main/java/io/github/nucleuspowered/nucleus/internal/annotations/SkipOnError.java
@@ -1,0 +1,17 @@
+/*
+ * This file is part of Nucleus, licensed under the MIT License (MIT). See the LICENSE.txt file
+ * at the root of this project for more details.
+ */
+package io.github.nucleuspowered.nucleus.internal.annotations;
+
+import java.lang.annotation.*;
+
+/**
+ * If the Guice injector is unable to create an instance of a class with this annotation, then Nucleus and QSML
+ * will not fail to load a module.
+ */
+@Target(ElementType.TYPE)
+@Retention(RetentionPolicy.RUNTIME)
+@Documented
+public @interface SkipOnError {
+}
