+++ src/main/java/io/github/nucleuspowered/nucleus/modules/core/listeners/ReloadListener.java
@@ -1,0 +1,27 @@
+/*
+ * This file is part of Nucleus, licensed under the MIT License (MIT). See the LICENSE.txt file
+ * at the root of this project for more details.
+ */
+package io.github.nucleuspowered.nucleus.modules.core.listeners;
+
+import io.github.nucleuspowered.nucleus.Util;
+import io.github.nucleuspowered.nucleus.internal.ListenerBase;
+import io.github.nucleuspowered.nucleus.internal.annotations.SkipOnError;
+import org.spongepowered.api.Sponge;
+import org.spongepowered.api.command.CommandSource;
+import org.spongepowered.api.event.Listener;
+import org.spongepowered.api.event.game.GameReloadEvent;
+import org.spongepowered.api.text.Text;
+import org.spongepowered.api.text.format.TextColors;
+
+@SkipOnError
+public class ReloadListener extends ListenerBase {
+
+    @Listener
+    public void onGameReload(final GameReloadEvent event) {
+        plugin.reload();
+        CommandSource requester = event.getCause().first(CommandSource.class).orElse(Sponge.getServer().getConsole());
+        requester.sendMessage(Text.of(TextColors.YELLOW, "[Nucleus] ", Util.getTextMessageWithFormat("command.reload.one")));
+        requester.sendMessage(Text.of(TextColors.YELLOW, "[Nucleus] ", Util.getTextMessageWithFormat("command.reload.two")));
+    }
+}
