+++ src/main/resources/assets/io/github/nucleuspowered/nucleus/messages.properties
@@ -33,6 +33,7 @@ startup.postinit={0} is now entering the post-init phase.
 startup.moduleloading={0} is now loading and enabling modules. This may take a few seconds.
 startup.modulenotloaded={0} was unable to load modules and has aborted loading.
 startup.moduleloaded={0} has completed loading modules.
+startup.injectablenotloaded=The {0} was not loaded because of an injection error, but loading will continue.
 startup.started={0} has started.
 startup.stopped={0} is performing server shutdown tasks.
